requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
textblob
newspaper3k
matplotlib
